Clean up constants package
